keep building with custom filter out from BaseRepository
