+++ src/Mado/QueryBundle/Repositories/BaseRepository.php
@@ -5,7 +5,6 @@ namespace Mado\QueryBundle\Repositories;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\QueryBuilder;
 use Hateoas\Representation\Factory\PagerfantaFactory;
-use Mado\QueryBundle\Exceptions\InvalidFiltersException;
 use Mado\QueryBundle\Objects\MetaDataAdapter;
 use Mado\QueryBundle\Objects\PagerfantaBuilder;
 use Mado\QueryBundle\Queries\QueryBuilderFactory;
@@ -49,7 +48,10 @@ class BaseRepository extends EntityRepository
         $this->metadata->setEntityName($this->getEntityName());
 
         $this->queryBuilderFactory = new QueryBuilderFactory($this->getEntityManager());
+
         $this->qoBuilder = new QueryOptionsBuilder();
+        $entityAlias = $this->metadata->getEntityAlias();
+        $this->qoBuilder->setEntityAlias($entityAlias);
     }
 
     public function initFromQueryBuilderOptions(QueryBuilderOptions $options)
@@ -99,71 +101,14 @@ class BaseRepository extends EntityRepository
 
     public function setQueryOptionsFromRequest(Request $request = null)
     {
-        $entityAlias = $this->metadata->getEntityAlias();
-        $this->qoBuilder->setEntityAlias($entityAlias);
         $this->queryOptions = $this->qoBuilder->builderFromRequest($request);
 
         return $this;
     }
 
-    private function ensureFilterIsValid($filters)
-    {
-        if (!is_array($filters)) {
-
-            $message = "Wrong query string exception: ";
-            $message .= var_export($filters, true) . "\n";
-            $message .= "Please check query string should be something like " .
-                "http://127.0.0.1:8000/?filtering[status]=todo";
-
-            throw new InvalidFiltersException($message);
-        }
-    }
-
     public function setQueryOptionsFromRequestWithCustomFilter(Request $request = null, $filter)
     {
-        $filters = $request->query->get('filtering', []);
-        $orFilters = $request->query->get('filtering_or', []);
-        $sorting = $request->query->get('sorting', []);
-        $printing = $request->query->get('printing', []);
-        $rel = $request->query->get('rel', '');
-        $page = $request->query->get('page', '');
-        $select = $request->query->get('select', $this->metadata->getEntityAlias());
-        $filtering = $request->query->get('filtering', '');
-        $limit = $request->query->get('limit', '');
-        $justCount = $request->query->get('justCount', 'false');
-
-        $this->ensureFilterIsValid($filters);
-        $filters = array_merge($filters, $filter);
-
-        $filterOrCorrected = [];
-
-        $count = 0;
-        foreach ($orFilters as $key => $filterValue) {
-            if (is_array($filterValue)) {
-                foreach ($filterValue as $keyInternal => $internal) {
-                    $filterOrCorrected[$keyInternal . '|' . $count] = $internal;
-                    $count += 1;
-                }
-            } else {
-                $filterOrCorrected[$key] = $filterValue;
-            }
-        }
-
-        $this->queryOptions = QueryBuilderOptions::fromArray([
-            '_route' => $request->attributes->get('_route'),
-            '_route_params' => $request->attributes->get('_route_params', []),
-            'id' => $request->attributes->get('id'),
-            'filtering' => $filtering,
-            'limit' => $limit,
-            'page' => $page,
-            'filters' => $filters,
-            'orFilters' => $filterOrCorrected,
-            'sorting' => $sorting,
-            'rel' => $rel,
-            'printing' => $printing,
-            'select' => $select,
-            'justCount' => $justCount,
-        ]);
+        $this->queryOptions = $this->qoBuilder->buildFromRequestAndCustomFilter($request, $filter);
 
         return $this;
     }
